+++ visualization/Readme.md
@@ -1,10 +1,10 @@
 This is a tool for data analysis. It has two pages: "Inference" and "Analyze".
 ## Inference page
-This page allows for the analysis of model answers depending on different parameters we set up. It has three modes: "Complete", "Run one test", and "Run whole dataset".
+This page allows for the analysis of model answers depending on different parameters we set up. It has three modes: "Chat", "Run one sample", and "Run whole dataset".
 
-**Complete** mode allows for a conversation with the model. It has the minimum parameters to set up.
+**Chat** mode allows for a conversation with the model. It has the minimum parameters to set up.
 
-**Run one test** mode allows you to send a single question to the model. It can be a question from the dataset (parameters `data_file` or `dataset` and `split_name` determine the dataset) or your custom question. The answer is also validated by comparing it with the `expected_answer` field.
+**Run one sample** mode allows you to send a single question to the model. It can be a question from the dataset (parameters `data_file` or `dataset` and `split_name` determine the dataset) or your custom question. The answer is also validated by comparing it with the `expected_answer` field.
 
 **Run whole dataset** mode allows you to launch the model with chosen parameters on the whole dataset. It will save results in `visualization/results/output-greedy.jsonl` and `visualization/results/metrics-greedy.jsonl`. If the "use random seed range" flag is turned on, the dataset will be launched for random seeds in the range from `start_random_seed` to `end_random_seed`. All results can then be analyzed on the "Analyze" page (you do not need to relaunch the program), and the parameters you used to launch the dataset can be found in the `visualization/results/parameters.jsonl` file and on the "Analyze" page as well.
