+++ visualization/layouts/run_prompt_page_layouts.py
@@ -17,12 +17,14 @@ from typing import List, Tuple
 import dash_bootstrap_components as dbc
 from dash import dcc, html
 from layouts.base_layouts import get_text_area_layout
-from settings.constants import COMPLETE_MODE, ONE_TEST_MODE, WHOLE_DATASET_MODE
+from settings.constants import CHAT_MODE, ONE_SAMPLE_MODE, WHOLE_DATASET_MODE
 from utils.common import get_examples
 from utils.strategies.strategy_maker import RunPromptStrategyMaker
 
 
-def get_few_shots_by_id_layout(page: int, examples_type: str, view_mode: bool) -> Tuple[html.Div]:
+def get_few_shots_by_id_layout(
+    page: int, examples_type: str, view_mode: bool
+) -> Tuple[html.Div]:
     examples_list = get_examples().get(
         examples_type,
         [{}],
@@ -46,11 +48,13 @@ def get_few_shots_by_id_layout(page: int, examples_type: str, view_mode: bool) -
 
 
 def get_query_params_layout(
-    mode: str = ONE_TEST_MODE,
+    mode: str = ONE_SAMPLE_MODE,
 ) -> List[dbc.AccordionItem]:
     strategy = RunPromptStrategyMaker(mode).get_strategy()
     return (
-        strategy.get_utils_input_layout() + strategy.get_few_shots_input_layout() + strategy.get_query_input_layout()
+        strategy.get_utils_input_layout()
+        + strategy.get_few_shots_input_layout()
+        + strategy.get_query_input_layout()
     )
 
 
@@ -64,11 +68,11 @@ def get_run_mode_layout() -> html.Div:
                 labelClassName="btn btn-outline-primary",
                 labelCheckedClassName="active",
                 options=[
-                    {"label": "Complete", "value": COMPLETE_MODE},
-                    {"label": "Run one test", "value": ONE_TEST_MODE},
+                    {"label": "Chat", "value": CHAT_MODE},
+                    {"label": "Run one sample", "value": ONE_SAMPLE_MODE},
                     {"label": "Run whole dataset", "value": WHOLE_DATASET_MODE},
                 ],
-                value=ONE_TEST_MODE,
+                value=ONE_SAMPLE_MODE,
             ),
         ],
         className="radio-group",
