+++ visualization/settings/constants.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 ANSWER_FIELD = "expected_answer"
-COMPLETE_MODE = "complete_mode"
+CHAT_MODE = "chat_mode"
 CHOOSE_MODEL = "choose_model"
 DATA_PAGE_SIZE = 10
 ERROR_MESSAGE_TEMPLATE = "When applying {} function\ngot errors\n{}"
@@ -21,7 +21,7 @@ GREEDY = "greedy"
 QUERY_INPUT_TYPE = "query_input"
 QUERY_INPUT_ID = '{{"type": "{}", "id": "{}"}}'
 QUESTION_FIELD = "question"
-ONE_TEST_MODE = "one_test"
+ONE_SAMPLE_MODE = "one_sample"
 METRICS = "metrics"
 OUTPUT = "output"
 OUTPUT_PATH = "{}-{}.jsonl"
