+++ visualization/utils/strategies/one_sample_mode.py
@@ -15,38 +15,38 @@
 from typing import Dict, List
 
 import dash_bootstrap_components as dbc
-from settings.constants import ANSWER_FIELD, ONE_TEST_MODE, QUESTION_FIELD
+from dash import html
+from settings.constants import ONE_SAMPLE_MODE, PARAMS_FOR_WHOLE_DATASET_ONLY
+from utils.common import get_test_data
 
 from visualization.utils.strategies.base_strategy import ModeStrategies
 
 
-class CompleteModeStrategy(ModeStrategies):
-    mode = ONE_TEST_MODE
+class OneTestModeStrategy(ModeStrategies):
+    mode = ONE_SAMPLE_MODE
 
     def __init__(self):
         super().__init__()
 
     def get_utils_input_layout(self) -> List[dbc.AccordionItem]:
-        inference_condition = lambda name, value: not isinstance(value, dict)
-
-        return super().get_utils_input_layout(inference_condition, lambda name, value: False, True)
-
-    def get_few_shots_input_layout(self) -> List[dbc.AccordionItem]:
-        return []
+        inference_condition = (
+            lambda name, value: isinstance(value, (str, int, float))
+            and name not in PARAMS_FOR_WHOLE_DATASET_ONLY
+        )
+        return super().get_utils_input_layout(
+            inference_condition,
+            lambda name, value: True,
+            True,
+            list(self.config.items()),
+        )
 
     def get_query_input_layout(self) -> List[dbc.AccordionItem]:
         return super().get_query_input_layout(
-            [
-                (QUESTION_FIELD, ""),
-                (ANSWER_FIELD, ""),
-            ],
-            False,
+            get_test_data(
+                0,
+            )[0].items()
         )
 
-    def run(self, utils: Dict, params: Dict):
-        utils['delimiter'] = "\n\n\n\n\n\n"
+    def run(self, utils: Dict, params: Dict) -> html.Div:
         params['prompts'] = [self.get_prompt(utils, params['question'])]
         return super().run(utils, params)
-
-    def get_prompt(self, utils: Dict, question: str) -> str:
-        return str(question)
