+++ visualization/utils/strategies/strategy_maker.py
@@ -14,9 +14,9 @@
 
 from typing import Optional
 
-from settings.constants import COMPLETE_MODE, ONE_TEST_MODE, WHOLE_DATASET_MODE
-from utils.strategies.complete_mode import CompleteModeStrategy
-from utils.strategies.one_test_mode import OneTestModeStrategy
+from settings.constants import CHAT_MODE, ONE_SAMPLE_MODE, WHOLE_DATASET_MODE
+from utils.strategies.chat_mode import ChatModeStrategy
+from utils.strategies.one_sample_mode import OneTestModeStrategy
 from utils.strategies.whole_dataset_mode import WholeDatasetModeStrategy
 
 from visualization.utils.strategies.base_strategy import ModeStrategies
@@ -24,9 +24,9 @@ from visualization.utils.strategies.base_strategy import ModeStrategies
 
 class RunPromptStrategyMaker:
     strategies = {
-        ONE_TEST_MODE: OneTestModeStrategy,
+        ONE_SAMPLE_MODE: OneTestModeStrategy,
         WHOLE_DATASET_MODE: WholeDatasetModeStrategy,
-        COMPLETE_MODE: CompleteModeStrategy,
+        CHAT_MODE: ChatModeStrategy,
     }
 
     def __init__(self, mode: Optional[str] = None):
